Refactored plugins section
