+++ src/engine/main.py
@@ -3,8 +3,8 @@ from pathlib import Path
 
 import fastapi
 from fastapi.responses import JSONResponse
-from .routers import test_plugins, package, deployment, plugins
-from .plugengine import PluginEngine
+from .routers import package, deployment, plugins
+from .engine import PluginEngine
 from . import __version__
 
 app = fastapi.FastAPI(
@@ -15,7 +15,6 @@ app = fastapi.FastAPI(
 )
 
 # Add the routes for all specialized routers
-app.include_router(test_plugins.router)
 app.include_router(package.router)
 app.include_router(deployment.router)
 app.include_router(plugins.router)
@@ -43,8 +42,8 @@ async def debug_exception_handler(request: fastapi.Request, exc: Exception):
 @app.on_event("startup")
 async def initialize_plugins():
     plugin_engine = PluginEngine()
-    plugin_engine._discover()
-    plugin_engine._load()
+    plugin_engine.discover()
+    plugin_engine.load()
 
 
 @app.get("/healthcheck")
