--- src/engine/plugengine/__init__.py
@@ -1,1 +1,0 @@
-from .engine import PluginEngine
